+++ src/components/Form/Fields/Checkbox/Checkbox.stories.tsx
@@ -5,7 +5,7 @@ import { useForm } from "react-hook-form";
 import { Checkbox } from "./Checkbox";
 import { Form } from "../../../Form/Form";
 import { ICheckbox } from "../../../../models/IFormInput";
-import { CustomSubmitButton } from "../../CustomSubmitButton";
+import { CustomSubmitButton } from "../../CustomSubmitButton/CustomSubmitButton";
 
 export default {
     title: "Form/Fields/Checkbox",